Changes to be committed: 	modified:   package-lock.json 	modified:   package.json 	modified:   src/app.css 	new file:   src/lib/all_page.ts 	new file:   src/lib/class/connection.ts 	modified:   src/lib/class/wifi.ts 	modified:   src/lib/components/Footer.svelte 	modified:   src/lib/components/Navigation.svelte 	modified:   src/lib/components/NotConnected.svelte 	modified:   src/lib/components/SetSchedule.svelte 	modified:   src/lib/components/ToastContainer.svelte 	new file:   src/lib/db.ts 	new file:   src/lib/enum.ts 	new file:   src/lib/messageStore.ts 	modified:   src/lib/modal/EmailVerification.svelte 	modified:   src/lib/modal/PlantModal.svelte 	modified:   src/lib/modal/ViewPicture.svelte 	modified:   src/lib/services/home/configService.ts 	modified:   src/lib/services/home/plantService.ts 	modified:   src/lib/services/home/robotService.ts 	new file:   src/lib/socket.ts 	modified:   src/lib/stores/connection.ts 	modified:   src/lib/stores/toast.ts 	modified:   src/lib/type.ts 	new file:   src/lib/utils/indexdb.ts 	modified:   src/lib/utils/request.ts 	new file:   src/routes/+error.svelte 	modified:   src/routes/+layout.svelte 	modified:   src/routes/+layout.ts 	deleted:    src/routes/+page.server.ts 	modified:   src/routes/+page.svelte 	new file:   src/routes/api/custom-sync/+server.ts 	new file:   src/routes/api/folder/+server.ts 	new file:   src/routes/api/sync/+server.ts 	modified:   src/routes/api/user/login/+server.ts 	modified:   src/routes/folder/[slug]/+page.server.ts 	modified:   src/routes/folder/[slug]/+page.svelte 	modified:   src/routes/live/+page.server.ts 	modified:   src/routes/live/+page.svelte 	new file:   src/routes/live/CameraFeed.svelte 	new file:   src/routes/live/LivestreamControls.svelte 	modified:   src/routes/login/+page.svelte 	modified:   src/routes/login/loginform.svelte 	modified:   src/routes/logs/+page.svelte 	deleted:    src/routes/notification/+page.server.ts 	modified:   src/routes/notification/+page.svelte 	deleted:    src/routes/record/+page.server.ts 	modified:   src/routes/record/+page.svelte 	modified:   src/routes/robot/+page.svelte 	modified:   src/routes/robot/driver.svelte 	modified:   src/routes/robot/head.svelte 	modified:   src/routes/robot/relay.svelte 	modified:   src/routes/robot/rgbsensor.svelte 	modified:   src/routes/robot/tcrt5000.svelte 	modified:   src/routes/robot/ultrasonic.svelte 	modified:   src/routes/robot/watersensor.svelte 	modified:   src/routes/spray/camera.svelte 	modified:   src/routes/spray/cameradesktop.svelte 	modified:   src/routes/spray/camerainfo.svelte 	modified:   src/routes/spray/cameramobile.svelte 	modified:   src/routes/spray/configuration.svelte 	modified:   src/routes/spray/details.svelte 	modified:   src/routes/spray/detectedplantlist.svelte 	modified:   src/routes/spray/scanbuttons.svelte 	modified:   src/routes/spray/scannedheader.svelte 	modified:   src/routes/spray/spraybuttons.svelte 	modified:   src/routes/spray/stoprobot.svelte 	new file:   src/routes/tailscale/+page.svelte 	deleted:    src/routes/talescale/+page.svelte 	new file:   src/routes/test/+page.svelte 	modified:   src/routes/wifi/+page.svelte 	modified:   src/routes/wifi/selectedssid.svelte 	modified:   src/routes/wifi/wifiheader.svelte 	modified:   src/routes/wifi/wifilist.svelte 	modified:   vite.config.ts

+++ src/lib/modal/ViewPicture.svelte
@@ -1,15 +1,18 @@
 <script lang="ts">
 	import { addToast, removeToast } from '$lib/stores/toast';
 	import { capitalize } from '$utils/string';
+	import type { PlantHistories } from '../type';
 
 	export let modalOpen;
 	export let closeModal: () => void;
 	export let downloadImage;
 	export let selectedImage;
-	export let images: any;
+	export let images: any | null;
 	export let noDownloadDelete = false;
 
 	async function deleteImage(public_id: string) {
+		if (images === null) return;
+
 		const toastId = addToast('Deleting image...', 'loading');
 		const res = await fetch('/api/delete-image', {
 			method: 'DELETE',
@@ -78,7 +81,7 @@
 					</p>
 					<p class="text-sm text-gray-600 dark:text-gray-400">
 						<strong>Plant Health:</strong>
-						{capitalize(selectedImage.diseaseName)}
+						{capitalize(selectedImage.plantHealth ?? selectedImage.diseaseName)}
 					</p>
 					<p class="text-sm text-gray-600 dark:text-gray-400">
 						<strong>Image Size:</strong>
